fastapi_microservice_with_tests/requirements.txt
fastapi~=0.115.12
uvicorn
pydantic~=2.11.3
python-jose[cryptography]
passlib[bcrypt]~=1.7.4
fastapi-pagination
pydantic[email]
python-multipart
bcrypt<4.0.0
